+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 - Création du module `Application`.
 - Intégration de GTK4.
 - Affichage de la première fenêtre.
+
+### Added
+
+- Module `MainWindow`.
+- Séparation de la fenêtre principale du module `Application`.
+- Première architecture de la fenêtre principale.